models/crypto_open_model_1/requirements.txt
numerapi===2.19.1
python-dotenv===1.0.0
pandas===2.1.4
pyarrow===14.0.2
xgboost===2.1.1
scikit-learn===1.3.2
psutil===5.9.7
numpy===1.26.2
yfinance===0.2.41